Implement @param tag handling
By parsing the route using Mustermann, we can find the named parameters; 
these can then be passed to the RouteObject so that YARD doesn't 
complain. It is important to set the docstring *after* the parameters 
are set, otherwise the unknown parameter warning will still occur once.

In RouteObject, we need to override the parameters attribute to prevent 
parameters from being shown in the signature of the route 'method', e.g. 
`#GET /:id(id)`, which looks strange.